+++ crates/web-client/test/transactions.test.ts
@@ -224,7 +224,8 @@ export const compileTxScript = async (
       true
     );
 
-    const compiledScript = await client.compileTxScript(_script);
+    let builder = new window.ScriptBuilder(window.ScriptBuilderMode.Debug);
+    const compiledScript = builder.compileNoteScript(_script);
 
     return {
       scriptRoot: compiledScript.root().toHex(),